Refactor `web_server/api_gateway` to fix circular import
A circular import was detected in the `web_server/api_gateway` module.
The `__init__.py` file was importing route modules, which were in turn
importing a Blueprint object from the `__init__.py` file.

This change resolves the issue by creating a new `gateway.py` file to
hold the Blueprint definition. The `__init__.py` and all route files
have been updated to import the Blueprint from this new file, breaking
the circular dependency.
