+++ web_server/api_gateway/document_routes.py
@@ -10,7 +10,7 @@ import os
 import tempfile
 import json
 from flask import request, jsonify, current_app
-from . import api_gateway_bp
+from .gateway import api_gateway_bp
 from .auth_middleware import token_required
 from ..app import db
 from web_server.models import Document, DocumentAnalysis
