+++ web_server/api_gateway/gateway.py
@@ -1,0 +1,8 @@
+"""
+API Gateway Blueprint
+"""
+
+from flask import Blueprint
+
+# Create API Gateway Blueprint
+api_gateway_bp = Blueprint('api_gateway', __name__)
