+++ web_server/api_gateway/organization_routes.py
@@ -6,7 +6,7 @@ Organization Management Routes for API Gateway
 
 import uuid
 from flask import request, jsonify, current_app
-from . import api_gateway_bp
+from .gateway import api_gateway_bp
 from .auth_middleware import token_required
 from ..app import db
 from ..models import Organization, OrganizationMember, User, Document, DocumentAnalysis
